BIG REFACTOR. Instead of a hardcoded struct (which was getting unwieldly) now the checkboxes are stored in an array much like the AppearanceStates. This allows me to iterate over and check/uncheck the boxes and show/hide them, making those corresponding functions MUCH shorter and more readable. My sanity is improving just looking at it.
UNDO now works for a single gender. (Mostly tested vs female up to now; more testing required.) Added a DLC_1 tracking flag to AppearanceState which finally fixed things; some cleanup required around this but right now it's functional, so it's a good time for a commit.

More cleanup and two new utility functions to support the new flag in AppearanceState.