FIXES BUG: When generating a new appearance, the new appearance isn't reflected in the vanilla UI.
This boiled down to calling UIScreen::UpdateData(), which up to now I'd only been doing for the gender change button. Also, instead of updating the camera after every
RandomizeTrait call (which I should've realized was expensive) I'm now doing it at the end of each appearance gen function (TotallyRandom + NormalAppearance). If Undo doesn't do this yet, it should.

Also bumped the buffer size up from 10 to 12 to see if that feels better.

Deleted vestigial Content.ini, which was created to see if I could fix the really annoying red screen error that came with the DLC. (It didn't work.)